Moved shared/samples/transfers templates into samples/transfers, updated tests
