+++ app/views/projects/samples/clones/_dialog.html.erb
@@ -10,7 +10,7 @@
       data-infinite-scroll-plural-description-value="<%= t(".description.plural") %>"
     >
       <%= form_with(
-        url: list_namespace_project_samples_path(list_class: 'sample'),
+        url: list_samples_path(list_class: 'sample'),
         data: { "infinite-scroll-target": "pageForm" }
       ) do %>
         <div data-infinite-scroll-target="pageFormContent"></div>
