+++ app/views/projects/samples/deletions/_new_multiple_deletions_dialog.html.erb
@@ -8,7 +8,7 @@
     data-infinite-scroll-plural-description-value="<%= t(".description.plural") %>"
   >
     <%= form_with(
-        url: list_namespace_project_samples_path(list_class: 'sample'),
+        url: list_samples_path(list_class: 'sample'),
         data: { "infinite-scroll-target": "pageForm" }
       ) do %>
       <div data-infinite-scroll-target="pageFormContent"></div>
