+++ app/views/samples/transfers/_dialog.html.erb
