+++ app/views/samples/transfers/new.turbo_stream.erb
@@ -1,6 +1,10 @@
-<%= render partial: "shared/samples/transfers/new",
-locals: {
-  list_samples_path: list_samples_path,
-  transfer_path: samples_transfer_path(namespace_id: @namespace.id),
-  broadcast_target: @broadcast_target,
-} %>
+<%= turbo_stream.update "samples_dialog",
+                    partial: "samples/transfers/dialog",
+                    locals: {
+                      open: true,
+                      closable: true,
+                      list_samples_path: list_samples_path,
+                      transfer_path:
+                        samples_transfer_path(namespace_id: @namespace.id),
+                      broadcast_target: @broadcast_target,
+                    } %>
