--- app/views/shared/samples/transfers/_create.turbo_stream.erb
@@ -1,22 +1,0 @@
-<% if Flipper.enabled?(:progress_bars) %>
-  <%= turbo_stream.update "samples_dialog",
-                      partial: "shared/samples/transfers/dialog",
-                      locals: {
-                        open: true,
-                        closable: false,
-                        list_samples_path: list_samples_path,
-                        transfer_path: transfer_path,
-                        broadcast_target: broadcast_target,
-                      } %>
-
-  <%= turbo_stream.update "transfer_samples_dialog_content" do %>
-    <%= render partial: "shared/progress_bar", locals: { percentage: 0 } %>
-  <% end %>
-
-<% else %>
-  <%= turbo_stream.append "transfer_samples_dialog_content" do %>
-    <%= render SpinnerComponent.new(
-      message: t("shared.samples.transfers.dialog.spinner_message"),
-    ) %>
-  <% end %>
-<% end %>
