--- app/views/shared/samples/transfers/_new.turbo_stream.erb
@@ -1,9 +1,0 @@
-<%= turbo_stream.update "samples_dialog",
-                    partial: "shared/samples/transfers/dialog",
-                    locals: {
-                      open: true,
-                      closable: true,
-                      list_samples_path: list_samples_path,
-                      transfer_path: transfer_path,
-                      broadcast_target: broadcast_target,
-                    } %>
